# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Api
 * AetherPulse API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserUpdateStatus } from './userUpdateStatus'

export interface UserUpdate {
  displayName?: string
  /**
   * @maxLength 190
   * @nullable
   */
  bio?: string | null
  /**
   * @maxLength 40
   * @nullable
   */
  pronouns?: string | null
  /**
   * @maxLength 120
   * @nullable
   */
  website?: string | null
  /**
   * @maxLength 40
   * @nullable
   */
  location?: string | null
  /**
   * @maxLength 128
   * @nullable
   */
  customStatus?: string | null
  /** @pattern ^#[0-9A-Fa-f]{6}$ */
  accentColor?: string
  /** @nullable */
  primaryColor?: string | null
  /** @nullable */
  displayNameStyle?: string | null
  /** @nullable */
  profileGradient?: string | null
  /** @nullable */
  avatarUrl?: string | null
  /** @nullable */
  bannerUrl?: string | null
  status?: UserUpdateStatus
  badges?: string[]
}