# shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AetherPulse API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          description: Forbidden
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, displayName, status, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ['string', 'null']
        bannerUrl:
          type: ['string', 'null']
        bio:
          type: ['string', 'null']
        pronouns:
          type: ['string', 'null']
        website:
          type: ['string', 'null']
        location:
          type: ['string', 'null']
        status:
          type: string
          enum: [online, away, busy, offline]
        customStatus:
          type: ['string', 'null']
        accentColor:
          type: ['string', 'null']
        primaryColor:
          type: ['string', 'null']
        displayNameStyle:
          type: ['string', 'null']
        profileGradient:
          type: ['string', 'null']
        avatarFrame:
          type: ['string', 'null']
        profileTheme:
          type: ['string', 'null']
        customTheme:
          type: ['string', 'null']
        preferredTheme:
          type: ['string', 'null']
        badges:
          type: array
          items:
            type: string
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        displayName:
          type: string
        bio:
          type: ['string', 'null']
          maxLength: 190
        pronouns:
          type: ['string', 'null']
          maxLength: 40
        website:
          type: ['string', 'null']
          maxLength: 120
        location:
          type: ['string', 'null']
          maxLength: 40
        customStatus:
          type: ['string', 'null']
          maxLength: 128
        accentColor:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        primaryColor:
          type: ['string', 'null']
        displayNameStyle:
          type: ['string', 'null']
        profileGradient:
          type: ['string', 'null']
        avatarFrame:
          type: ['string', 'null']
        profileTheme:
          type: ['string', 'null']
        customTheme:
          type: ['string', 'null']
        preferredTheme:
          type: ['string', 'null']
        avatarUrl:
          type: ['string', 'null']
        bannerUrl:
          type: ['string', 'null']
        status:
          type: string
          enum: [online, away, busy, offline]
        badges:
          type: array
          items:
            type: string
...